Allowing usage of NyxxRest in web, (by disabling forced FFI on unsupported platforms)
