+++ lib/src/ffi_real.dart
@@ -1,0 +1,9 @@
+export 'dart:ffi';
+import 'dart:io';
+
+/// The name of the zstd library on the current platform.
+final zstdLibraryName = Platform.isWindows
+    ? 'zstd.dll'
+    : Platform.isMacOS
+        ? 'libzstd.dylib'
+        : 'libzstd.so';
