+++ lib/src/ffi_stub.dart
@@ -1,0 +1,15 @@
+/// A stub for the dart:ffi library.
+class DynamicLibrary {
+  /// A stub for DynamicLibrary.open.
+  static DynamicLibrary open(String path) {
+    throw UnsupportedError('DynamicLibrary is not supported on this platform.');
+  }
+
+  /// A stub for DynamicLibrary.close.
+  void close() {
+    throw UnsupportedError('DynamicLibrary is not supported on this platform.');
+  }
+}
+
+/// A stub for the zstd library name.
+const zstdLibraryName = '';
