+++ lib/src/gateway/decompress_zstd_stub.dart
@@ -1,0 +1,3 @@
+Stream<dynamic> decompressZStdTransport(Stream<List<int>> raw) {
+  throw UnsupportedError('ZStd decompression is not supported on this platform.');
+}
